Make NullTranslator return ICU's neutral locale
ICU does not really have the notion of a neutral locale.
Also, the "C" locale has no special meaning there.

The closest thing to POSIX's "C" locale is "en_US_POSIX"
(american english with POSIX-compatible semantics)
We use that because it is expected that this library will only
be used with classes/functions related to ICU.